refactor(macos): extract Home carousels from AppModel into AppModel+Home.swift
Continues the AppModel god-object split. Relocates the Home-screen carousel
loaders — Jump Back In, Recently Added, Quick Picks, Favorite Albums / Artists,
Rediscover, Recently Discovered Artists, Suggestions, plus shuffle / show-all-
favorites — into a new `extension AppModel` file.

Pure relocation: the 184 moved lines are byte-identical to what was removed.
AppModel.swift drops 4,650 -> 4,466 lines. The carousel result arrays stay as
@Observable stored state on the main class.

Four access promotions (private -> internal): the items-query fetch helpers the
moved carousels call — fetchAlbumsViaItemsQuery, fetchAlbumsWithPlayCounts,
fetchLatestAlbumsWithDates, fetchFavoriteTracks — kept under a new "Items query
helpers" MARK.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
